Fix supervised model and evaluator so the dimension are correctly set

+++ tests/test_cstr_kmc_model.f90
@@ -79,7 +79,7 @@ Module test_cstr_kmc_model
 Contains
   Subroutine test_cstr_kmc_parameters_getNParameters(this, nPar, nCstr)
     class(test_cstr_kmc_parameters), intent(in), target :: this
-    integer(8), intent(out) :: nPar, nCstr
+    integer, intent(out) :: nPar, nCstr
     nPar = 5
     nCstr = 0
   End Subroutine test_cstr_kmc_parameters_getNParameters
